Added convenience function boolToString. Improved compile time assertions. Removed assertf() in favor of panic(). Improved IRQ, ISR, PIC code readability. Added debug logging to multiboot2 header parser. Moved regs structure to isrs.h

+++ src/kernel/system.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <system.h>
 
-uint32 intDivCeil(uint32 a, uint32 b) { // calculates ceil(a/b) without using any floating point math
+uint64 intDivCeil(uint64 a, uint64 b) { // calculates ceil(a/b) without using any floating point math
 	if (a % b == 0) {
 		return a / b;
 	} else {
